api/requirements.txt
Flask==2.2.2
Jinja2==3.1.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
python-dotenv==0.21.0
PyJWT==2.6.0
Pillow==9.2.0
mongodb-migrations==1.2.1
flask-cors
requests
cryptography
pycryptodome